wikipedia-fetcher API
Developed a platform to query over wikipedia. API can help to get basic information for any query. Results can be
served in two formats as JSON as well as HTML format. Spring MVC was used for serving in the backend.
Backend process involves calling wikipedia.org to fetch raw data, which is eventually parsed and cleaned to get the
output. Tested the project for 300+ queries like “India”, “America”, “Einstein” etc.